+++ orb/emailer.py
@@ -5,6 +5,30 @@ from django.core.urlresolvers import reverse
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext as _
 
+
+def user_welcome(to_user):
+    template_html = 'orb/email/welcome.html'
+    template_text = 'orb/email/welcome.txt'
+    
+    from_email = settings.SERVER_EMAIL
+    subject = settings.EMAIL_SUBJECT_PREFIX + _(u"Welcome to ORB")
+    
+    data = {"firstname": to_user.first_name,
+            "lastname": to_user.last_name,
+            "info_email": settings.ORB_INFO_EMAIL }
+    
+    text_content = render_to_string(template_text, data)
+    html_content = render_to_string(template_html, data)
+
+    send_mail(subject, 
+              text_content, 
+              from_email,
+              [to_user.email], 
+              fail_silently=False, 
+              html_message=html_content)
+            
+    return 
+
 def password_reset(to_user, new_password):
     template_html = 'orb/email/password_reset.html'
     template_text = 'orb/email/password_reset.txt'