+++ orb/signals.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 from django.db import models
 from django.dispatch import Signal
 
-from orb.emailer import first_resource, resource_approved, resource_rejected
+from orb.emailer import first_resource, resource_approved, resource_rejected, user_welcome
 from orb.models import ResourceTracker, SearchTracker, ResourceWorkflowTracker
 from orb.lib.search_crawler import is_search_crawler
 
@@ -14,12 +14,13 @@ resource_workflow = Signal(providing_args=["request", "resource", "status", "not
 resource_url_viewed = Signal(providing_args=["resource_url", "request"])
 resource_file_viewed = Signal(providing_args=["resource_file", "request"])
 search = Signal(providing_args=["query", "no_results", "request"])
-user_registered = Signal(providing_args=["user"])
+user_registered = Signal(providing_args=["user", "request"])
 
-def user_registered(sender, **kwargs):
-    if sended.created:
-        pass
-    pass
+def user_registered_callback(sender, **kwargs):
+    request = kwargs.get('request')
+    user = kwargs.get('user')
+    user_welcome(user)
+    return
 
 def resource_viewed_callback(sender, **kwargs):
     request = kwargs.get('request')
@@ -52,12 +53,12 @@ def resource_workflow_callback(sender, **kwargs):
     if status == ResourceWorkflowTracker.PENDING_CRT:
         no_previous_resources = ResourceWorkflowTracker.objects.filter(create_user=request.user).count()
         if no_previous_resources == 0:
-            #first_resource(request.user, resource)
+            first_resource(request.user, resource)
             email_sent = True
     
     # if approved
     if status == ResourceWorkflowTracker.APPROVED:
-        #resource_approved(request, resource.create_user, resource)
+        resource_approved(request, resource.create_user, resource)
         email_sent = True
     
     # if passed to MEP
@@ -66,7 +67,7 @@ def resource_workflow_callback(sender, **kwargs):
     
     # if rejected
     if status == ResourceWorkflowTracker.REJECTED:
-        #resource_rejected(resource.create_user, resource, notes)
+        resource_rejected(resource.create_user, resource, notes)
         email_sent = True
         pass
             
@@ -139,3 +140,4 @@ resource_workflow.connect(resource_workflow_callback)
 resource_url_viewed.connect(resource_url_viewed_callback)
 resource_file_viewed.connect(resource_file_viewed_callback)
 search.connect(search_callback)
+user_registered.connect(user_registered_callback)